feat: add community recognized badge to market views
