+++ packages/kit/src/views/Market/MarketDetailV2/components/TokenDetailHeader/TokenDetailHeaderLeft.tsx
@@ -1,20 +1,16 @@
-import { useIntl } from 'react-intl';
-
 import {
   Divider,
-  Icon,
   InteractiveIcon,
   SizableText,
-  Tooltip,
   XStack,
   YStack,
 } from '@onekeyhq/components';
 import { Token } from '@onekeyhq/kit/src/components/Token';
-import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { EWatchlistFrom } from '@onekeyhq/shared/src/logger/scopes/dex';
 import accountUtils from '@onekeyhq/shared/src/utils/accountUtils';
 import type { IMarketTokenDetail } from '@onekeyhq/shared/types/marketV2';
 
+import { CommunityRecognizedBadge } from '../../../components/CommunityRecognizedBadge';
 import { MarketStarV2 } from '../../../components/MarketStarV2';
 import { TokenSecurityAlert } from '../TokenSecurityAlert';
 
@@ -36,7 +32,6 @@ export function TokenDetailHeaderLeft({
   showMediaAndSecurity = true,
   isNative = false,
 }: ITokenDetailHeaderLeftProps) {
-  const intl = useIntl();
   const {
     handleCopyAddress,
     handleOpenContractAddress,
@@ -85,25 +80,7 @@ export function TokenDetailHeaderLeft({
           <SizableText size="$bodyLgMedium" color="$text">
             {symbol}
           </SizableText>
-          {communityRecognized ? (
-            <Tooltip
-              placement="top"
-              renderTrigger={
-                <Icon
-                  name="BadgeRecognizedSolid"
-                  size="$4"
-                  color="$iconSuccess"
-                />
-              }
-              renderContent={
-                <SizableText size="$bodySm">
-                  {intl.formatMessage({
-                    id: ETranslations.dexmarket_communityRecognized,
-                  })}
-                </SizableText>
-              }
-            />
-          ) : null}
+          {communityRecognized ? <CommunityRecognizedBadge /> : null}
         </XStack>
 
         <XStack gap="$2" ai="center">
